refactor(naming): rename phase4.py to composition.py for naming consistency
- Create composition.py as new home for CompositionValidator
- Replace phase4.py with deprecation stub that emits module-level warning
- Rename Phase4ValidatorConfig → CompositionValidatorConfig with alias
- Rename test files to match new naming convention:
  - test_phase4_integration.py → test_composition_integration.py
  - test_phase4_validator.py → test_composition_validator_compat.py
  - test_phase4_additional.py → test_composition_validator_additional.py
- Update all internal imports to use .composition module
- Update documentation (validation.rst, README.md, migration guide)
- Add module deprecation test to verify warning is emitted

All backward compatibility maintained via deprecation stubs and aliases.
