+++ tests/unit/test_composition_validator.py
@@ -4,10 +4,13 @@ This module tests:
 - CompositionValidator can be imported and instantiated
 - Phase4Validator still works as a deprecated alias
 - Deprecation warning is emitted when using Phase4Validator
+- Module-level deprecation warning for phase4.py module
 """
 
 from __future__ import annotations
 
+import importlib
+import sys
 import warnings
 
 
@@ -117,3 +120,58 @@ class TestCompositionValidatorFunctionality:
             matrix_size=21,
         )
         assert callable(getattr(validator, "validate_centerpiece_safety", None))
+
+
+class TestPhase4ModuleDeprecation:
+    """Test that importing from phase4 module emits deprecation warning."""
+
+    def test_phase4_module_import_emits_warning(self) -> None:
+        """Test that importing from phase4 module emits deprecation warning."""
+        # Remove phase4 from sys.modules to ensure fresh import
+        modules_to_remove = [key for key in sys.modules.keys() if "phase4" in key]
+        for mod in modules_to_remove:
+            del sys.modules[mod]
+
+        with warnings.catch_warnings(record=True) as w:
+            warnings.simplefilter("always")
+
+            # Fresh import should trigger module-level deprecation
+            importlib.import_module("segnomms.validation.phase4")
+
+            # Check that a deprecation warning was issued about the module
+            module_deprecation_warnings = [
+                warning
+                for warning in w
+                if issubclass(warning.category, DeprecationWarning)
+                and "phase4" in str(warning.message).lower()
+                and "module" in str(warning.message).lower()
+            ]
+            assert len(module_deprecation_warnings) >= 1, (
+                f"Expected module deprecation warning for phase4, "
+                f"got warnings: {[str(x.message) for x in w]}"
+            )
+
+    def test_composition_module_no_deprecation_warning(self) -> None:
+        """Test that importing from composition module does NOT emit warning."""
+        # Remove composition from sys.modules to ensure fresh import
+        modules_to_remove = [key for key in sys.modules.keys() if "composition" in key]
+        for mod in modules_to_remove:
+            del sys.modules[mod]
+
+        with warnings.catch_warnings(record=True) as w:
+            warnings.simplefilter("always")
+
+            # Fresh import of composition should NOT trigger warning
+            importlib.import_module("segnomms.validation.composition")
+
+            # Check that no deprecation warning was issued about composition
+            composition_warnings = [
+                warning
+                for warning in w
+                if issubclass(warning.category, DeprecationWarning)
+                and "composition" in str(warning.message).lower()
+            ]
+            assert len(composition_warnings) == 0, (
+                f"Did not expect deprecation warning for composition, "
+                f"got warnings: {[str(x.message) for x in w]}"
+            )
